test: relocate main::tests
